добавил в массив СуперСобаку ограничив дистанции

+++ SuperCat.java
@@ -2,7 +2,7 @@ package GBTheme4oopRazbor;
 
 public class SuperCat extends Cat {
 
-    public SuperCat(String name,int runLimit,int jumpLimit) {
+    public SuperCat(String name,int runLimit,double jumpLimit) {
         super(name);
         this.jumpLimit=jumpLimit;
         this.runLimit=runLimit;
